DetailsMetrics Code Optimization
