+++ src/common/Chart/HistogramChart/histogramChart.scss
