+++ src/components/DetailsDrillDownAlert/DetailsAlertsMetrics.js
@@ -22,10 +22,10 @@ import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import { isEmpty } from 'lodash'
 
-import ApplicationMetricCard from '../DetailsMetrics/ApplicationMetricCard'
+import ApplicationMetricCard from '../DetailsMetrics/MetricsCards/ApplicationMetricCard'
 import DatePicker from '../../common/DatePicker/DatePicker'
 import NoData from '../../common/NoData/NoData'
-import NoMetricData from '../DetailsMetrics/NoMetricData'
+import NoMetricData from '../DetailsMetrics/MetricsCards/NoMetricData'
 import StatsCard from '../../common/StatsCard/StatsCard'
 
 import { REQUEST_CANCELED } from '../../constants'
